bigdecimal

+++ src/main/java/com/stockexchange/orderservice/model/Order.java
@@ -2,6 +2,7 @@ package com.stockexchange.orderservice.model;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -13,13 +14,13 @@ public class Order {
     private String symbol;
     private OrderType type;
     private OrderStatus status;
-    private double price;
+    private BigDecimal price;
     private int executedQuantity;
     private int totalQuantity;
     private Instant createdAt;
     private UUID userId;
 
-    public Order(UUID id, String symbol, OrderType type, OrderStatus status, double price, int executedQuantity, int totalQuantity, Instant createdAt, UUID userId) {
+    public Order(UUID id, String symbol, OrderType type, OrderStatus status, BigDecimal price, int executedQuantity, int totalQuantity, Instant createdAt, UUID userId) {
         this.orderId = id;
         this.symbol = symbol;
         this.type = type;
@@ -41,7 +42,7 @@ public class Order {
         return type;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice(){
         return price;
     }
